Revert "Restructured app with Flask Blueprints"
