+++ flaskapp/models.py
@@ -1,6 +1,4 @@
-
 from datetime import datetime
-from flask import current_app 
 from flaskapp import db, login_manager
 from flask_login import UserMixin
