+++ flaskapp/templates/layout.html
@@ -31,14 +31,14 @@
                 <!-- Set links -->
                 <div class="collapse navbar-collapse" id="navbarToggle">
                     <div class="navbar-nav mr-auto">    
-                        <a class="nav-item nav-link" href="{{ url_for('main.home') }}">Home</a>
-                        <a class="nav-item nav-link" href="{{ url_for('main.about') }}">About</a>
+                        <a class="nav-item nav-link" href="{{ url_for('home') }}">Home</a>
+                        <a class="nav-item nav-link" href="{{ url_for('about') }}">About</a>
                     </div>
 
                     <!-- Navbar Right Side: Updated per session status -->
                     <div class="navbar-nav">
                         {% if current_user.is_authenticated %}
-                            <a class="btn btn-primary btn-sml" href="{{ url_for('projects.new_project') }}">Create Project</a>
+                            <a class="btn btn-primary btn-sml" href="{{ url_for('new_project') }}">Create Project</a>
                             <!-- <button class="btn success">Success</button> -->
 
                             <!-- Profile pic drop down menu -->
@@ -47,12 +47,12 @@
                                     <img src="{{ url_for('static', filename='profile_pics/' + current_user.profile_image) }}" width="30" height="30" class="rounded-circle">
                                 </a>
                                 <div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
-                                    <a class="dropdown-item" href="{{ url_for('users.account') }}">Account</a>
-                                    <a class="dropdown-item" href="{{ url_for('users.logout') }}">Logout</a>
+                                    <a class="dropdown-item" href="{{ url_for('account') }}">Account</a>
+                                    <a class="dropdown-item" href="{{ url_for('logout') }}">Logout</a>
                                 </div>
                         {% else %}
-                            <a class="nav-item nav-link" href="{{ url_for('users.login') }}">Login</a>
-                            <a class="nav-item nav-link" href="{{ url_for('users.register') }}">Register</a>
+                            <a class="nav-item nav-link" href="{{ url_for('login') }}">Login</a>
+                            <a class="nav-item nav-link" href="{{ url_for('register') }}">Register</a>
                         {% endif %}
                     </div>
                 </div>
