+++ flaskapp/templates/project.html
@@ -15,7 +15,7 @@
                 <!-- Update/Delete buttons if project owner updates -->
                 {% if current_project.owner == current_user %}
                     
-                    <a class="btn btn-secondary btn-sm mt-1 mb-1" href="{{ url_for('projects.update_project', project_id=current_project.id) }}">Update</a>
+                    <a class="btn btn-secondary btn-sm mt-1 mb-1" href="{{ url_for('update_project', project_id=current_project.id) }}">Update</a>
                     <button type="button" class="btn btn-danger btn-sm m-1" data-toggle="modal" data-target="#deleteModal">Delete</button>
                     
 
@@ -23,10 +23,10 @@
                 {% else %}
                     <!-- Show join button -->
                     {% if current_user not in current_project.members  %}
-                        <a class="btn btn-primary btn-sm mt-1 mb-1" href="{{ url_for('projects.join_project', project_id=current_project.id) }}">Join</a>
+                        <a class="btn btn-primary btn-sm mt-1 mb-1" href="{{ url_for('join_project', project_id=current_project.id) }}">Join</a>
                     <!-- if already joined show leave button -->
                     {% else  %}
-                        <a class="btn btn-danger btn-sm mt-1 mb-1" href="{{ url_for('projects.leave_project', project_id=current_project.id) }}"> Leave </a>
+                        <a class="btn btn-danger btn-sm mt-1 mb-1" href="{{ url_for('leave_project', project_id=current_project.id) }}"> Leave </a>
                     {% endif %}
                     
                 {% endif %}
@@ -67,7 +67,7 @@
                 <div class="modal-footer">
                 <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
                 
-                <form action="{{ url_for('projects.delete_project', project_id=current_project.id) }}" method="POST">
+                <form action="{{ url_for('delete_project', project_id=current_project.id) }}" method="POST">
                     <input class="btn btn-danger" type="submit" value="Delete">
                 </form>
                 </div>
