+++ flaskapp/templates/register.html
@@ -101,7 +101,7 @@
 <div class="border-top pt-3">
 	<small class="text-muted">
 		Already Have An Account?
-		<a class="ml-2" href="{{ url_for('users.login') }}">Sign In</a>
+		<a class="ml-2" href="{{ url_for('login') }}">Sign In</a>
 	</small>
 </div>
 {% endblock content %}
