+++ run.py
@@ -1,10 +1,6 @@
-from flaskapp import create_app
-
-from flaskapp import db
+from flaskapp import app, db
 from flaskapp.tests.db_tables import bootstrap_helper, query_helper
 
-app = create_app()
-app.app_context().push()
 
 @app.cli.command("db-clean")
 def reset_db():
@@ -19,7 +15,7 @@ def bootstrap_data():
     ''' Populates database with some sample data '''
     db.drop_all()
     db.create_all()
-    bootstrap_helper(db,app)
+    bootstrap_helper(db)
     print("Loading data...\n")
     print("Re-Initialized and Bootstrapped with sample data")
 
@@ -27,9 +23,8 @@ def bootstrap_data():
 @app.cli.command("db-query")
 def query_sampleData():
     ''' Queries sample data from `boot` command '''
-    with app.app_context():
-        query_helper(db,app)
-        print("Querying Sample Data...")
+    query_helper(db)
+    print("Querying Sample Data...")
 
 
 if __name__ == '__main__':
